+++ modules/indexing/src/test/java/org/apache/ignite/internal/processors/cache/IgniteCacheAbstractQuerySelfTest.java
@@ -475,7 +475,7 @@ public abstract class IgniteCacheAbstractQuerySelfTest extends GridCommonAbstrac
             Cache.Entry<BinaryObject, BinaryObject> entry = F.first(qry.getAll());
 
             assertNotNull(entry);
-            assertEquals(100500L, entry.getKey().field("id"));
+            assertEquals(100500L, (long)entry.getKey().field("id"));
             assertEquals(val1, entry.getValue().deserialize());
         }
     }
